Navbar highlight threasholds now dynamic depending on page location
Navbar highlight threashold used to be static region in the middle 50% of the viewport. This means it could not physically highlight the contact us section.
Navbar highlight now sets the threashold to the top/bottom 50% of the page depending on if you where at the top/bottom scroll limit of the page respectivly.